requirements.txt
# === DEPENDENCIAS DEL CLASIFICADOR INTELIGENTE ===
# Nota: Si usas Gráfica NVIDIA y necesitas aceleración CUDA, instala PyTorch con:
# pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.7
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
idna==3.11
invoke==3.0.3
Jinja2==3.1.6
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
ml_dtypes==0.5.4
mpmath==1.3.0
ncnn==1.0.20260114
networkx==3.4.2
numpy==2.2.6
onnx==1.21.0
opencv-python==4.13.0.92
packaging==26.0
paramiko==4.0.0
pillow==12.1.1
pnnx==20260409
polars==1.39.3
polars-runtime-32==1.39.3
portalocker==3.2.0
protobuf==7.34.1
psutil==7.2.2
pycparser==3.0
PyNaCl==1.6.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pywin32==311
PyYAML==6.0.3
requests==2.33.1
scipy==1.15.3
scp==0.15.0
six==1.17.0
sympy==1.14.0
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.3
typing_extensions==4.15.0
ultralytics==8.4.37
ultralytics-thop==2.0.18
urllib3==2.6.3
